refactor LTCC reconstruction viewer
