+++ reconstruction/ltcc/src/main/java/org/jlab/service/ltcc/viewer/LTCCClusterHistos.java
@@ -3,8 +3,10 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package org.jlab.service.ltcc;
+package org.jlab.service.ltcc.viewer;
+import org.jlab.service.ltcc.viewer.LTCCHistogrammer;
 import org.jlab.groot.data.H1F;
+import org.jlab.service.ltcc.LTCCCluster;
 
 
 /**
